requirements.txt
python-dateutil>=2.0
django-jsonfield>=0.9.2
django-qsstats-magic>=0.7.2
django-cache-utils>=0.7.2
django-admin-tools>=0.5.1
django-nvd3>=0.5.0
django-bower
